Phase 2: Create folder architecture and move files
