+++ src/components/providers/FirebaseAuthBridge.tsx
