+++ src/features/groups/components/FilterBar.tsx
